feature: notification library
feature: veiled layouts
feature: stricter control on uploads

+++ app/src/main/java/za/co/xisystems/itis_rrm/utils/enums/ToastDuration.kt
@@ -2,11 +2,11 @@ package za.co.xisystems.itis_rrm.utils.enums
 
 import www.sanju.motiontoast.MotionToast
 
-enum class ToastDuration(private val value: Int) {
+enum class ToastDuration(private val value: Long) {
     LONG(MotionToast.LONG_DURATION),
     SHORT(MotionToast.SHORT_DURATION);
 
-    fun getValue(): Int {
+    fun getValue(): Long {
         return value
     }
 }
